added reviews

+++ app/src/main/java/com/handaoui/movies/data/Comment.kt
@@ -5,6 +5,6 @@ import java.util.*
 data class Comment(
         val commentator: String,
         val content: String,
-        val rating: Int = -1,
+        val rating: Float = -1f,
         val date: Date = Date()
 )
